requirements.txt
discord.py==2.4.0
rasa==3.6.20